fix: auth 처리 페이지 분리
